requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
fastapi
uvicorn
pypdf
ollama
streamlit
requests